Add concerns and aedg_import id mapping model
This is a test impelementation to see how we might implement tracking
id's from the AEDG data team when importing the data into rails. This
way we can make sure we're safe from issues when a foreign ID value
might not match the internal database handling of auto incrementing IDs.

Signed-off-by: Will Fisher <whfisher@alaska.edu>