Remove the default db:reset when doing an import all
When doing the initial setup of a rails application, we don't need to
delete everything as there shouldn't be anything in the
db yet and rails will complain if it's running in production mode.

We also already catch when data is detected in the tables during import
and the user can elect to clear everything with a `db:reset` or the
given table delete command.

Signed-off-by: Will Fisher <whfisher@alaska.edu>